use portable data types from stdint.h

+++ cnpy.cpp
@@ -8,6 +8,7 @@
 #include<algorithm>
 #include<cstring>
 #include<iomanip>
+#include<stdint.h>
 
 char cnpy::BigEndianTest() {
     int x = 1;
@@ -98,7 +99,7 @@ void cnpy::parse_npy_header(FILE* fp, size_t& word_size, size_t*& shape, size_t&
     word_size = atoi(str_ws.substr(0,loc2).c_str());
 }
 
-void cnpy::parse_zip_footer(FILE* fp, unsigned short& nrecs, size_t& global_header_size, size_t& global_header_offset)
+void cnpy::parse_zip_footer(FILE* fp, uint16_t& nrecs, size_t& global_header_size, size_t& global_header_offset)
 {
     std::vector<char> footer(22);
     fseek(fp,-22,SEEK_END);
@@ -106,14 +107,14 @@ void cnpy::parse_zip_footer(FILE* fp, unsigned short& nrecs, size_t& global_head
     if(res != 22)
         throw std::runtime_error("parse_zip_footer: failed fread");
 
-    unsigned short disk_no, disk_start, nrecs_on_disk, comment_len;
-    disk_no = *(unsigned short*) &footer[4];
-    disk_start = *(unsigned short*) &footer[6];
-    nrecs_on_disk = *(unsigned short*) &footer[8];
-    nrecs = *(unsigned short*) &footer[10];
-    global_header_size = *(unsigned int*) &footer[12];
-    global_header_offset = *(unsigned int*) &footer[16];
-    comment_len = *(unsigned short*) &footer[20];
+    uint16_t disk_no, disk_start, nrecs_on_disk, comment_len;
+    disk_no = *(uint16_t*) &footer[4];
+    disk_start = *(uint16_t*) &footer[6];
+    nrecs_on_disk = *(uint16_t*) &footer[8];
+    nrecs = *(uint16_t*) &footer[10];
+    global_header_size = *(uint32_t*) &footer[12];
+    global_header_offset = *(uint32_t*) &footer[16];
+    comment_len = *(uint16_t*) &footer[20];
 
     assert(disk_no == 0);
     assert(disk_start == 0);
@@ -126,7 +127,7 @@ cnpy::NpyArray load_the_npy_file(FILE* fp) {
     size_t ndims, word_size;
     bool fortran_order;
     cnpy::parse_npy_header(fp,word_size,shape,ndims,fortran_order);
-    unsigned long long size = 1; //long long so no overflow when multiplying by word_size
+    size_t size = 1; //long long so no overflow when multiplying by word_size
     for(size_t i = 0;i < ndims;i++) size *= shape[i];
 
     cnpy::NpyArray arr;
@@ -158,7 +159,7 @@ cnpy::npz_t cnpy::npz_load(std::string fname) {
         if(local_header[2] != 0x03 || local_header[3] != 0x04) break;
 
         //read in the variable name
-        unsigned short name_len = *(unsigned short*) &local_header[26];
+        uint16_t name_len = *(uint16_t*) &local_header[26];
         std::string varname(name_len,' ');
         size_t vname_res = fread(&varname[0],sizeof(char),name_len,fp);
         if(vname_res != name_len)
@@ -168,7 +169,7 @@ cnpy::npz_t cnpy::npz_load(std::string fname) {
         varname.erase(varname.end()-4,varname.end());
 
         //read in the extra field
-        unsigned short extra_field_len = *(unsigned short*) &local_header[28];
+        uint16_t extra_field_len = *(uint16_t*) &local_header[28];
         if(extra_field_len > 0) {
             std::vector<char> buff(extra_field_len);
             size_t efield_res = fread(&buff[0],sizeof(char),extra_field_len,fp);
@@ -201,7 +202,7 @@ cnpy::NpyArray cnpy::npz_load(std::string fname, std::string varname) {
         if(local_header[2] != 0x03 || local_header[3] != 0x04) break;
 
         //read in the variable name
-        unsigned short name_len = *(unsigned short*) &local_header[26];
+        uint16_t name_len = *(uint16_t*) &local_header[26];
         std::string vname(name_len,' ');
         size_t vname_res = fread(&vname[0],sizeof(char),name_len,fp);      
         if(vname_res != name_len)
@@ -209,7 +210,7 @@ cnpy::NpyArray cnpy::npz_load(std::string fname, std::string varname) {
         vname.erase(vname.end()-4,vname.end()); //erase the lagging .npy
 
         //read in the extra field
-        unsigned short extra_field_len = *(unsigned short*) &local_header[28];
+        uint16_t extra_field_len = *(uint16_t*) &local_header[28];
         fseek(fp,extra_field_len,SEEK_CUR); //skip past the extra field
 
         if(vname == varname) {
@@ -219,7 +220,7 @@ cnpy::NpyArray cnpy::npz_load(std::string fname, std::string varname) {
         }
         else {
             //skip past the data
-            size_t size = *(unsigned int*) &local_header[22];
+            size_t size = *(uint32_t*) &local_header[22];
             fseek(fp,size,SEEK_CUR);
         }
     }
